+++ src/app/modules/community/community.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { TwitterComponent } from './twitter/twitter.component';
-import { StackoverflowComponent } from './stackoverflow/stackoverflow.component';
+import { TwitterComponent } from './components/twitter/twitter.component';
+import { StackoverflowComponent } from './components/stackoverflow/stackoverflow.component';
 import { CommunitySummaryComponent } from './shared/community-summary/community-summary.component';
 import { AppMaterialModule } from '../../app-material.module';
