+++ src/app/modules/community/components/stackoverflow/stackoverflow.component.html
@@ -1,4 +1,4 @@
-<div *ngFor="let post of stackoverflowPosts">
+<div *ngFor="let post of stackOverflowPosts">
   <mat-card class="mb-2">
     <mat-card-content>
       <p>{{post.title}}</p>
