+++ src/app/modules/community/models/stack-overflow-post.ts
@@ -1,0 +1,4 @@
+export interface StackOverflowPost {
+  title: string;
+  link: string;
+}
