+++ src/app/modules/community/services/community.service.spec.ts
@@ -1,15 +1,15 @@
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { TestBed } from '@angular/core/testing';
 
-import { StackoverflowService } from './stackoverflow.service';
+import { CommunityService } from './community.service';
 
-describe('StackoverflowService', () => {
+describe('CommunityService', () => {
   beforeEach(() => TestBed.configureTestingModule({
     imports: [HttpClientTestingModule]
   }));
 
   it('should be created', () => {
-    const service: StackoverflowService = TestBed.get(StackoverflowService);
+    const service: CommunityService = TestBed.get(CommunityService);
     expect(service).toBeTruthy();
   });
 });
