--- src/app/modules/community/stackoverflow/stackoverflow-post.ts
@@ -1,4 +1,0 @@
-export interface StackoverflowPost {
-  title: string;
-  link: string;
-}
