+++ src/app/modules/projects/project.module.ts
@@ -1,7 +1,7 @@
 import { AppMaterialModule } from './../../app-material.module';
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { ProjectSummaryComponent } from './shared/project-summary/project-summary.component';
+import { ProjectSummaryComponent } from './shared/components/project-summary/project-summary.component';
 
 @NgModule({
   declarations: [ProjectSummaryComponent],
