+++ src/app/modules/projects/shared/components/project-summary/project-summary.component.spec.ts
